@@ -4,7 +4,7 @@ description: Demonstrates how to use the volume plugin.
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ^1.10.0
 
 
