@@ -5,7 +5,7 @@ version: 1.0.1
 homepage: https://github.com/aman-malhotra-1999/Volume_Flutter.git
 
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ^1.10.0
 dependencies:
   flutter:
